+++ recipe/key/laravel.php
@@ -3,4 +3,7 @@
 namespace Deployer;
 
 require_once 'recipe/laravel.php';
+
+set('key_platform', 'LARAVEL');
+
 require_once __DIR__ . '/../key.php';